+++ oscm-ui-tests/src/main/resources/webtest.properties
@@ -21,4 +21,5 @@ technicalservice.xml.path=/build/oscm-ui-tests/src/main/resources/TechnicalServi
 technicalservice.name=AppSampleService
 create.subscription.count=10
 create.subscription.waiting.seconds=10
-auth.mode=OSCM authentication mode (INTERNAL or OIDC)
+#auth.mode=OSCM authentication mode (INTERNAL or OIDC)
+auth.mode=INTERNAL
