+++ oscm-ui-tests/src/test/java/portal/PortalMarketServiceWT.java
@@ -8,7 +8,13 @@
 
 package portal;
 
-import org.junit.*;
+import static org.junit.Assert.assertTrue;
+
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
+import org.junit.FixMethodOrder;
+import org.junit.Rule;
+import org.junit.Test;
 import org.junit.rules.TestWatcher;
 import org.junit.runners.MethodSorters;
 import org.openqa.selenium.By;
@@ -116,7 +122,6 @@ public class PortalMarketServiceWT {
         tester.waitForElementVisible(
                 By.id(PortalHtmlElements.DEFINE_PRICEMODEL_BUTTON_SAVE), 10);
         tester.clickElement(PortalHtmlElements.DEFINE_PRICEMODEL_BUTTON_SAVE);
-
         assertTrue(tester.getExecutionResult());
 
     }
