+++ oscm-ui-tests/src/test/java/portal/PortalTechServiceWT.java
@@ -8,7 +8,13 @@
 
 package portal;
 
-import org.junit.*;
+import static org.junit.Assert.assertTrue;
+
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
+import org.junit.FixMethodOrder;
+import org.junit.Rule;
+import org.junit.Test;
 import org.junit.rules.TestWatcher;
 import org.junit.runners.MethodSorters;
 import org.openqa.selenium.By;
@@ -16,8 +22,6 @@ import org.openqa.selenium.support.ui.Select;
 import webtest.PortalHtmlElements;
 import webtest.PortalPathSegments;
 import webtest.PortalTester;
-
-import static org.junit.Assert.assertTrue;
 
 /**
  * Integration web test to create a technical service.
